Fix position sizing and add real account balance fetching
- Fixed open order retrieval by using Info client instead of Exchange client
- Fixed position sizing to properly cap at  max risk per trade
- Added get_account_balance() method to fetch real portfolio value
- Updated strategy manager to use real account balance instead of hardcoded
- Fixed position limit check to use actual capital at risk instead of position count
- Added periodic account balance updates during trading
- Fixed ORDER_TIMEOUT_MINUTES parsing to handle float values

+++ src/config/settings.py
@@ -44,7 +44,7 @@ def load_config() -> Dict[str, Any]:
             "stop_loss_percentage": float(os.getenv("STOP_LOSS_PERCENTAGE", "2.0")),
             
             # Order monitoring settings
-            "order_timeout_minutes": int(os.getenv("ORDER_TIMEOUT_MINUTES", "30")),
+            "order_timeout_minutes": float(os.getenv("ORDER_TIMEOUT_MINUTES", "30")),
             "enable_stale_order_cleanup": os.getenv("ENABLE_STALE_ORDER_CLEANUP", "true").lower() == "true",
             "position_sync_interval": int(os.getenv("POSITION_SYNC_INTERVAL", "60")),  # seconds
             "enable_position_validation": os.getenv("ENABLE_POSITION_VALIDATION", "true").lower() == "true",
